fix: correcting from merge dosen
- only route (dosen), not backend
